gestao_senhas/requirements.txt
Django==4.2.0
python-decouple==3.8
Pillow>=12.0.0
dj-database-url==2.1.0
PyMySQL==1.1.0
